requirements.txt
streamlit==1.29.0
openai-whisper==20231117
torch==2.1.0
transformers==4.35.0
datasets==2.14.0
gtts==2.4.0
sounddevice==0.4.6
numpy==1.24.3
pandas==2.1.3
requests==2.31.0


fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.354
langchain-openai==0.0.2
langchain-community==0.0.10
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
python-dotenv==1.0.0

